working scripts and openzwave

+++ src/Database.cpp
@@ -148,7 +148,7 @@ namespace micasa {
 	// The above template is specialized for the types listed below.
 	template int Database::getQueryValue( const std::string& query_, ... ) const;
 	template unsigned int Database::getQueryValue( const std::string& query_, ... ) const;
-	template float Database::getQueryValue( const std::string& query_, ... ) const;
+	template double Database::getQueryValue( const std::string& query_, ... ) const;
 
 	// The string variant of the above template doesn't require string streams and has it's own
 	// specialized implementation.
